Check if the column is there before trying to remove.

+++ db/migrate/20120130205050_remove_photos_count_from_post.rb
@@ -1,17 +1,13 @@
 class RemovePhotosCountFromPost < ActiveRecord::Migration
   def self.up
-    remove_column :posts, :photos_count
+    # Not all DBs will have this, depending on when they started running Diaspora
+    # Re: historical migrations, schemas, etc.
+    if column_exists?( :posts, :photos_count )
+      remove_column :posts, :photos_count
+    end
   end
 
   def self.down
-    add_column :posts, :photos_count, :integer, :default => 0
-    execute %{
-      UPDATE posts
-      SET photos_count = (
-        SELECT COUNT(*)
-        FROM photos
-        WHERE photos.status_message_guid = posts.guid
-      )
-    }
+    # The column should never have been there.
   end
 end
